clear warning

+++ cpymo/cpymo_game_selector.c
@@ -255,7 +255,7 @@ static error_t cpymo_game_selector_lazy_init_update(cpymo_engine *e, void *ui_, 
 
 		int relative = 0;
 		if (last_selected) {
-			while (last_selected->prev && relative < data->nodes_per_screen - 1) {
+			while (last_selected->prev && relative < (int)data->nodes_per_screen - 1) {
 				last_selected = last_selected->prev;
 				relative++;
 			}
